refactor: Game 클래스 controller 패키지 생성하여 이동
