+++ src/main/java/baseball/Main.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.controller.Game;
+
 public class Main {
 
     public static void main(String[] args) {
